VB-650 Refactor Integration Test and Unit tests, removed extra dir.. move code back
