+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/config/ClientTrackingConfigurationTest.kt
@@ -16,7 +16,7 @@ import org.springframework.mock.web.MockHttpServletResponse
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.context.junit.jupiter.SpringExtension
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.JwtAuthHelper
+import uk.gov.justice.digital.hmpps.visitscheduler.helper.JwtAuthHelper
 
 @Import(JwtAuthHelper::class, ClientTrackingInterceptor::class, ClientTrackingConfiguration::class)
 @ContextConfiguration(initializers = [ConfigDataApplicationContextInitializer::class])
