+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/OpenApiDocsTest.kt
@@ -2,7 +2,6 @@ package uk.gov.justice.digital.hmpps.visitscheduler.integration
 
 import org.junit.jupiter.api.Test
 import org.springframework.http.MediaType
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.integration.IntegrationTestBase
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
