+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/PrisonApiMockServer.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.visitscheduler.test_setup.integration
+package uk.gov.justice.digital.hmpps.visitscheduler.integration
 
 import com.github.tomakehurst.wiremock.WireMockServer
 import com.github.tomakehurst.wiremock.client.WireMock.aResponse
