+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/VisitSessionTemplateControllerTest.kt
@@ -13,11 +13,10 @@ import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency
 import uk.gov.justice.digital.hmpps.visitscheduler.model.SessionFrequency.WEEKLY
 import uk.gov.justice.digital.hmpps.visitscheduler.model.VisitType
 import uk.gov.justice.digital.hmpps.visitscheduler.repository.SessionTemplateRepository
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.TestClockConfiguration
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplate
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplateCreator
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.helper.sessionTemplateDeleter
-import uk.gov.justice.digital.hmpps.visitscheduler.test_setup.integration.IntegrationTestBase
+import uk.gov.justice.digital.hmpps.visitscheduler.helper.TestClockConfiguration
+import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplate
+import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplateCreator
+import uk.gov.justice.digital.hmpps.visitscheduler.helper.sessionTemplateDeleter
 import java.time.LocalDate
 import java.time.LocalTime
